+++ src/main/java/com/parse/gwt/client/ParsePointer.java
@@ -1,4 +1,4 @@
-package com.dotweblabs.gwt.client;
+package com.parse.gwt.client;
 
 import com.google.gwt.json.client.JSONObject;
 import com.google.gwt.json.client.JSONString;